refactor: useChatMessages 훅 리팩토링 및 날짜 구분 유틸 분리
